pyproject.toml
[project]
name = "lying-deception-experiments"
version = "0.1.0"
description = "Multi-agent experiments on lying aversion vs deception aversion in economic organizations"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
    "scikit-learn>=1.2",
]

[project.scripts]
lying-exp = "main:main"
